tester/requirements.txt
psycopg[binary]==3.2.3
httpx==0.27.2
pyyaml==6.0.2
apscheduler==3.10.4
tenacity==9.0.0
